requirements.txt
flask
sqlalchemy
flask-login
flask-wtf
flask-mail
flask-bcrypt
flask-cors
python-dotenv
alembic